Reorder deletion operations (#21)
This is so that when the directory deletion operations take too long,
they are still applied in the next run. This works because the users
being present on the system signals to the script that the user still
needs to be deleted.

The motivation of this PR is that this workflow run timed out, and on
the next run, the directories are not deleted.
https://github.com/WATonomous/infra-config/actions/runs/14687944657/job/41219376118